pyproject.toml
[project]
name = "gameonix-dc"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "discord-py>=2.5.2",
    "dotenv>=0.9.9",
    "psutil>=7.0.0",
    "pymongo>=4.12.1",
]
